requirements.txt
asgiref==3.6.0
Django==4.1.5
sqlparse==0.4.3
tzdata==2022.7
black==22.12.0
chardet==5.1.0
colorama==0.4.6
djangorestframework==3.14.0